First draft of testing
Currently forces user to have function application happen inside of the
library so that we can control execution and avoid propagating
exceptions to top-level or having a function not terminate (instead of
timing out).

Requires a description string for each test and a toString function for
the output type of the function being tested.

Not sure if it is worth it to monomoprhise to avoid toString / just have
a shared ref and count up for test #s to avoid the description string.